Fix all merge conflicts around settings.go
